extract support lib
