+++ ansys/mapdl/reader/_version.py
@@ -8,7 +8,7 @@ version_info = 0, 58, 'dev0'
 """
 
 # major, minor, patch
-version_info = 0, 52, 'dev0'
+version_info = 0, 53, 'dev0'
 
 # Nice string for the version
 __version__ = '.'.join(map(str, version_info))
